+++ packages/terra-framework-docs/CHANGELOG.md
@@ -4,7 +4,20 @@
 * Added
   * Added examples and tests for `terra-worklist-data-grid` to interact with cells containing focusable elements.
 
-## 1.26.0 - (July 20, 2023)
+## 1.30.0 - (August 2, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 1.29.0 - (August 2, 2023)
+
+* Added
+  * Added tests for draggable tabs.
+
+## 1.28.0 - (July 31, 2023)
+
+* Changed
+  * Minor dependency version bump
 
 * Added
   * Added examples for `terra-worklist-data-grid` when using pinned columns.
@@ -19,6 +32,7 @@
 
 * Added
   * Added `terra-worklist-data-grid` as dependency.
+  * Added example to test toggling and focusing disclosing node for `terra-slide-panel`.
 
 ## 1.25.0 - (July 19, 2023)