+++ packages/terra-navigation-prompt/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 1.75.0 - (August 2, 2023)
+
+* Changed
+  * Locked `uuid` dependency to `3.4.0` for consistency across Terra packages.
+
+## 1.74.0 - (July 31, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 1.73.0 - (July 25, 2023)
 
 * Changed