+++ packages/terra-theme-provider/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 4.13.0 - (July 31, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 4.12.0 - (February 14, 2023)
 
 * Changed